Merge pull request !23 from 把红色赋予黑海/SCAPI1.9

+++ Engine/Engine/Window.cs
@@ -580,7 +580,7 @@ namespace Engine {
             if (m_state != State.Uncreated) {
                 Display.Resize();
 #if !BROWSER
-                Scale = m_view.Size.X > 0f ? m_view.FramebufferSize.X / m_view.Size.X : 1f;
+                Scale = m_view.Size.X > 0 ? (float)m_view.FramebufferSize.X / m_view.Size.X : 1f;
 #endif
                 Resized?.Invoke();
             }
